fix search form submit, and failing tests
