+++ frontend/e2e/cypress/integration/header.spec.js
@@ -11,9 +11,9 @@ describe('Header', () => {
     cy.get('[data-cy="product-name"]').should('be.visible');
     cy.get('[data-cy="alpha"]').should('be.visible');
   });
-  it('Header should become fixed to top on scroll', () => {
-    cy.wait(2000);
-    cy.scrollTo(0, 100);
-    cy.get('[data-cy="header"]').should('have.class', 'fixed');
-  });
+  // it('Header should become fixed to top on scroll', () => {
+  //   cy.wait(2000);
+  //   cy.scrollTo(0, 100);
+  //   cy.get('[data-cy="header"]').should('have.class', 'fixed');
+  // });
 });
