+++ frontend/src/components/buttons/Close.tsx
@@ -10,6 +10,7 @@ const Close = ({ onClick, size = '20' }: CloseProps) => {
     <button
       className="focus:outline-none pointer-events-auto"
       onClick={onClick}
+      type="button"
       style={{
         height: `${size}px`,
         width: `${size}px`,
