+++ frontend/src/components/buttons/SearchButton.tsx
@@ -1,6 +1,9 @@
 const SearchButton = ({ children }) => {
   return (
-    <button className="bg-indigo-600 text-white uppercase p-2 rounded-full hover:bg-blue-500 transtion duration-300 w-16 md:w-24 shrink-0 font-medium">
+    <button
+      type="submit"
+      className="bg-indigo-600 text-white uppercase p-2 rounded-full hover:bg-blue-500 transtion duration-300 w-16 md:w-24 shrink-0 font-medium"
+    >
       {children}
     </button>
   );
