+++ frontend/src/pages/index.tsx
@@ -1,6 +1,6 @@
 import Layout from '../components/layouts/FullPageBanner';
 import Dashboard from '../components/Dashboard';
-import SearchInput from '../components/form-inputs/SearchInput';
+import SearchForm from '../components/forms/SearchForm';
 
 const IndexPage = () => {
   return (
@@ -11,7 +11,7 @@ const IndexPage = () => {
       <div className="absolute top-0 left-0 w-full z-10 mt-52 md:mt-44 ">
         <div className="container py-4 overflow-x-hidden">
           <Dashboard />
-          <SearchInput />
+          <SearchForm />
         </div>
       </div>
     </Layout>
